[REFACTOR] 리그 조회 로직 분리 #86
