+++ src/main/java/com/sports/server/query/repository/LeagueSportQueryRepository.java
@@ -1,11 +1,12 @@
-package com.sports.server.command.league.domain;
+package com.sports.server.query.repository;
 
+import com.sports.server.command.league.domain.LeagueSport;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.repository.Repository;
 
 import java.util.List;
 
-public interface LeagueSportRepository extends Repository<LeagueSport, Long> {
+public interface LeagueSportQueryRepository extends Repository<LeagueSport, Long> {
 
     @EntityGraph(attributePaths = "sport")
     List<LeagueSport> findByLeagueId(Long leagueId);
